requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg[binary]==3.1.12
sqlalchemy==2.0.23
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
httpx==0.25.2
apscheduler==3.10.4
python-dotenv==1.0.0
pyyaml==6.0.1
jinja2==3.1.2
